+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#193] - Shortcuts moving to previous/next order (thanks [@Xyz39808])
 - Cursor jump (Go To)
 - [#193] - Shortcut to select previous/next instrument (thanks [@Xyz39808])
+- Visible selection for toolbars and status bar
 - 2 new sample modules (thanks [@RigidatoMS] and [@ImATrackMan])
 
 ### Changed